move and rename files to new architecture
