+++ tests/Unit/inc/Engine/CDN/Subscriber/getCdnHosts.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace WP_Rocket\Tests\Unit\inc\classes\subscriber\CDN\CDNSubscriber;
+namespace WP_Rocket\Tests\Unit\inc\Engine\CDN\Subscriber;
 
 use Brain\Monkey\Functions;
 use Mockery;
 use WPMedia\PHPUnit\Unit\TestCase;
 use WP_Rocket\Admin\Options_Data;
-use WP_Rocket\CDN\CDN;
-use WP_Rocket\Subscriber\CDN\CDNSubscriber;
+use WP_Rocket\Engine\CDN\CDN;
+use WP_Rocket\Engine\CDN\Subscriber;
 
 /**
- * @covers \WP_Rocket\Subscriber\CDN\CDNSubscriber::get_cdn_hosts
+ * @covers \WP_Rocket\Engine\CDN\Subscriber::get_cdn_hosts
  * @group  CDN
  */
 class Test_GetCdnHosts extends TestCase {
@@ -21,7 +21,7 @@ class Test_GetCdnHosts extends TestCase {
 		parent::setUp();
 
 		$this->cdn        = Mockery::mock( CDN::class );
-		$this->subscriber = new CDNSubscriber(
+		$this->subscriber = new Subscriber(
 			Mockery::mock( Options_Data::class ),
 			$this->cdn
 		);
